Use millisecond precision for `timestamp` and `sentAt` fields (#42)
Add milliseconds to the timestamp and sentAt fields in order to preserve event chronology and avoid precision errors.

The timestamp field gets modified when Segment receives it, in order to correct for device clock skew. The documentation about timestamp explains the calculation:

    Segment calculates timestamp as timestamp = receivedAt - (sentAt - originalTimeStamp).

In cases where timestamp and sentAt are not the same (passing a custom timestamp, or the background job taking some time to execute and bumping sentAt to the next second), the lack of precision can cause the chronology of events to be affected.

Even if chronology of events is not affected, the events can have an error of up to 1 second in their timestamp.

This should be a non-breaking change. If some users are sending a seconds-precision timestamp this will add .000 and Segment will treat it exactly the same.

Just in case it helps, the official gem also added ms precision: segmentio/analytics-ruby#67 (in the time_in_iso8601 method)